add test_oai to openai workflow
